requirements.txt
streamlit
pandas
azure-ai-textanalytics